requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-core
langchain-ollama
faiss-cpu
sentence-transformers
unstructured
python-multipart
pypdf
pandas
openpyxl
langchain-huggingface
langchain-community
docx2txt
python-docx
langchain_ollama